sql_continue

+++ src/main/java/ru/hogwarts/school/conrollers/StudentController.java
@@ -52,7 +52,7 @@ public class StudentController {
     }
 
     @GetMapping("/get-average")
-    public ResponseEntity<Integer> getAverage() {
+    public ResponseEntity<Double> getAverage() {
         return ResponseEntity.ok(studentService.getAverageAgeFromStudents());
     }
